OAK-2709 - Misleading log message from IndexCopier
Made deletion of file more robust and exposed various stats as part of JMX  to get an insight into how much garbage is getting collected and how much is being accumulated

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1671489 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-lucene/src/main/java/org/apache/jackrabbit/oak/plugins/index/lucene/IndexCopier.java
@@ -22,11 +22,14 @@ package org.apache.jackrabbit.oak.plugins.index.lucene;
 import java.io.File;
 import java.io.IOException;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.Executor;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
@@ -40,6 +43,7 @@ import javax.management.openmbean.TabularDataSupport;
 import javax.management.openmbean.TabularType;
 
 import com.google.common.base.Charsets;
+import com.google.common.base.Function;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
@@ -58,10 +62,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static com.google.common.base.Preconditions.checkState;
+import static com.google.common.collect.Iterables.toArray;
+import static com.google.common.collect.Iterables.transform;
 import static com.google.common.collect.Maps.newConcurrentMap;
 
 class IndexCopier implements CopyOnReadStatsMBean {
     private static final Set<String> REMOTE_ONLY = ImmutableSet.of("segments.gen");
+    private static final int MAX_FAILURE_ENTRIES = 10000;
 
     private final Logger log = LoggerFactory.getLogger(getClass());
     private final Executor executor;
@@ -70,11 +77,21 @@ class IndexCopier implements CopyOnReadStatsMBean {
     private final AtomicInteger localReadCount = new AtomicInteger();
     private final AtomicInteger remoteReadCount = new AtomicInteger();
     private final AtomicInteger invalidFileCount = new AtomicInteger();
+    private final AtomicInteger deletedFileCount = new AtomicInteger();
+    private final AtomicInteger scheduledForCopyCount = new AtomicInteger();
+    private final AtomicInteger copyInProgressCount = new AtomicInteger();
+    private final AtomicInteger maxCopyInProgressCount = new AtomicInteger();
+    private final AtomicInteger maxScheduledForCopyCount = new AtomicInteger();
+    private final AtomicLong copyInProgressSize = new AtomicLong();
     private final AtomicLong downloadSize = new AtomicLong();
+    private final AtomicLong garbageCollectedSize = new AtomicLong();
     private final AtomicLong downloadTime = new AtomicLong();
+
 
     private final Map<String, String> indexPathMapping = Maps.newConcurrentMap();
     private final Map<String, String> indexPathVersionMapping = Maps.newConcurrentMap();
+    private final ConcurrentMap<String, LocalIndexFile> failedToDeleteFiles = Maps.newConcurrentMap();
+    private final Set<LocalIndexFile> copyInProgressFiles = Collections.newSetFromMap(new ConcurrentHashMap<LocalIndexFile, Boolean>());
 
     public IndexCopier(Executor executor, File indexRootDir) {
         this.executor = executor;
@@ -106,6 +123,36 @@ class IndexCopier implements CopyOnReadStatsMBean {
     public File getIndexDir(String indexPath) {
         String subDir = Hashing.sha256().hashString(indexPath, Charsets.UTF_8).toString();
         return new File(indexRootDir, subDir);
+    }
+
+    Map<String, LocalIndexFile> getFailedToDeleteFiles() {
+        return Collections.unmodifiableMap(failedToDeleteFiles);
+    }
+
+    private void failedToDelete(LocalIndexFile file){
+        //Limit the size on best effort basis
+        if (failedToDeleteFiles.size() < MAX_FAILURE_ENTRIES) {
+            LocalIndexFile failedToDeleteFile = failedToDeleteFiles.putIfAbsent(file.getKey(), file);
+            if (failedToDeleteFile == null){
+                failedToDeleteFile = file;
+            }
+            failedToDeleteFile.incrementAttemptToDelete();
+        } else {
+            log.warn("Not able to delete {}. Currently more than {} file with total size {} are pending delete.",
+                    file.deleteLog(), failedToDeleteFiles.size(), getGarbageSize());
+        }
+    }
+
+    private void successfullyDeleted(LocalIndexFile file, boolean fileExisted){
+        LocalIndexFile failedToDeleteFile = failedToDeleteFiles.remove(file.getKey());
+        if (failedToDeleteFile != null){
+            log.debug("Deleted : {}", failedToDeleteFile.deleteLog());
+        }
+
+        if (fileExisted){
+            garbageCollectedSize.addAndGet(file.size);
+            deletedFileCount.incrementAndGet();
+        }
     }
 
     /**
@@ -184,6 +231,7 @@ class IndexCopier implements CopyOnReadStatsMBean {
         }
 
         private void copy(final FileReference reference) {
+            updateMaxScheduled(scheduledForCopyCount.incrementAndGet());
             executor.execute(new Runnable() {
                 @Override
                 public void run() {
@@ -191,13 +239,17 @@ class IndexCopier implements CopyOnReadStatsMBean {
                     boolean success = false;
                     boolean copyAttempted = false;
                     try {
+                        scheduledForCopyCount.decrementAndGet();
                         if (!local.fileExists(name)) {
-                            long start = System.currentTimeMillis();
+                            long fileSize = remote.fileLength(name);
+                            LocalIndexFile file = new LocalIndexFile(local, name, fileSize);
+                            long start = startCopy(file);
                             copyAttempted = true;
+
                             remote.copy(local, name, name, IOContext.READ);
                             reference.markValid();
-                            downloadTime.addAndGet(System.currentTimeMillis() - start);
-                            downloadSize.addAndGet(remote.fileLength(name));
+
+                            doneCopy(file, start);
                         } else {
                             long localLength = local.fileLength(name);
                             long remoteLength = remote.fileLength(name);
@@ -289,17 +341,21 @@ class IndexCopier implements CopyOnReadStatsMBean {
             Set<String> failedToDelete = Sets.newHashSet();
 
             for (String fileName : filesToBeDeleted) {
+                LocalIndexFile file = new LocalIndexFile(local, fileName);
                 try {
-                    local.deleteFile(fileName);
+                    boolean fileExisted = false;
+                    if (local.fileExists(fileName)) {
+                        fileExisted = true;
+                        local.deleteFile(fileName);
+                    }
+                    successfullyDeleted(file, fileExisted);
                 } catch (IOException e) {
                     failedToDelete.add(fileName);
-                    log.debug("Error occurred while removing deleted file {} from Local {} ", fileName, local, e);
+                    failedToDelete(file);
+                    log.debug("Error occurred while removing deleted file {} from Local {}. " +
+                            "Attempt would be maid to delete it on next run ", fileName, local, e);
                 }
             }
-
-            log.info("Error occurred while deleting following files from the local index directory [{}]. " +
-                    "This can happen on Windows based system. Attempt would be made to remove them " +
-                    "in next attempt ", local, failedToDelete);
 
             filesToBeDeleted = new HashSet<String>(filesToBeDeleted);
             filesToBeDeleted.removeAll(failedToDelete);
@@ -332,6 +388,40 @@ class IndexCopier implements CopyOnReadStatsMBean {
         }
     }
 
+    private long startCopy(LocalIndexFile file) {
+        updateMaxInProgress(copyInProgressCount.incrementAndGet());
+        copyInProgressSize.addAndGet(file.size);
+        copyInProgressFiles.add(file);
+        return System.currentTimeMillis();
+    }
+
+    private void doneCopy(LocalIndexFile file, long start) {
+        copyInProgressFiles.remove(file);
+        copyInProgressCount.decrementAndGet();
+        copyInProgressSize.addAndGet(-file.size);
+
+        downloadTime.addAndGet(System.currentTimeMillis() - start);
+        downloadSize.addAndGet(file.size);
+    }
+
+    private void updateMaxScheduled(int val) {
+        synchronized (maxScheduledForCopyCount){
+            int current = maxScheduledForCopyCount.get();
+            if (val > current){
+                maxScheduledForCopyCount.set(val);
+            }
+        }
+    }
+
+    private void updateMaxInProgress(int val) {
+        synchronized (maxCopyInProgressCount){
+            int current = maxCopyInProgressCount.get();
+            if (val > current){
+                maxCopyInProgressCount.set(val);
+            }
+        }
+    }
+
     private class DeleteOldDirOnClose extends FilterDirectory {
         private final File oldIndexDir;
 
@@ -351,6 +441,107 @@ class IndexCopier implements CopyOnReadStatsMBean {
             super.close();
         }
     }
+    
+    static final class LocalIndexFile {
+        final File dir;
+        final String name;
+        final long size;
+        private volatile int deleteAttemptCount;
+        final long creationTime = System.currentTimeMillis();
+        
+        public LocalIndexFile(Directory dir, String fileName, long size){
+            this.dir = getFSDir(dir);
+            this.name = fileName;
+            this.size = size;
+        }
+
+        public LocalIndexFile(Directory dir, String fileName){
+            this(dir, fileName, getFileLength(dir, fileName));
+        }
+
+        public String getKey(){
+            if (dir != null){
+                return new File(dir, name).getAbsolutePath();
+            }
+            return name;
+        }
+
+        public void incrementAttemptToDelete(){
+            deleteAttemptCount++;
+        }
+
+        public int getDeleteAttemptCount() {
+            return deleteAttemptCount;
+        }
+
+        public String deleteLog(){
+            return String.format("%s (%s, %d attempts, %d s)", name,
+                    IOUtils.humanReadableByteCount(size), deleteAttemptCount, timeTaken());
+        }
+
+        public String copyLog(){
+            return String.format("%s (%s, %1.1f%%, %s, %d s)", name,
+                    IOUtils.humanReadableByteCount(actualSize()),
+                    copyProgress(),
+                    IOUtils.humanReadableByteCount(size), timeTaken());
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+
+            LocalIndexFile localIndexFile = (LocalIndexFile) o;
+
+            if (dir != null ? !dir.equals(localIndexFile.dir) : localIndexFile.dir != null)
+                return false;
+            return name.equals(localIndexFile.name);
+
+        }
+
+        @Override
+        public int hashCode() {
+            int result = dir != null ? dir.hashCode() : 0;
+            result = 31 * result + name.hashCode();
+            return result;
+        }
+
+        private long timeTaken(){
+            return TimeUnit.MILLISECONDS.toSeconds(System.currentTimeMillis() - creationTime);
+        }
+
+        private float copyProgress(){
+            return actualSize() * 1.0f / size * 100;
+        }
+
+        private long actualSize(){
+            return dir != null ? new File(dir, name).length() : 0;
+        }
+    }
+
+    static File getFSDir(Directory dir) {
+        if (dir instanceof FilterDirectory){
+            dir = ((FilterDirectory) dir).getDelegate();
+        }
+
+        if (dir instanceof FSDirectory){
+            return ((FSDirectory) dir).getDirectory();
+        }
+
+        return null;
+    }
+
+    /**
+     * Get the file length in best effort basis.
+     * @return actual fileLength. -1 if cannot determine
+     */
+    private static long getFileLength(Directory dir, String fileName){
+        try{
+            return dir.fileLength(fileName);
+        } catch (Exception e){
+            return -1;
+        }
+    }
 
     //~------------------------------------------< CopyOnReadStatsMBean >
 
@@ -362,9 +553,10 @@ class IndexCopier implements CopyOnReadStatsMBean {
                     "Lucene Index Stats", IndexMappingData.TYPE, new String[]{"jcrPath"});
             tds = new TabularDataSupport(tt);
             for (Map.Entry<String, String> e : indexPathMapping.entrySet()){
+                String size = IOUtils.humanReadableByteCount(FileUtils.sizeOfDirectory(new File(e.getValue())));
                 tds.put(new CompositeDataSupport(IndexMappingData.TYPE,
                         IndexMappingData.FIELD_NAMES,
-                        new String[] {e.getKey(), e.getValue()}));
+                        new String[]{e.getKey(), e.getValue(), size}));
             }
         } catch (OpenDataException e){
             throw new IllegalStateException(e);
@@ -401,18 +593,87 @@ class IndexCopier implements CopyOnReadStatsMBean {
         return IOUtils.humanReadableByteCount(FileUtils.sizeOfDirectory(indexRootDir));
     }
 
+    @Override
+    public String[] getGarbageDetails() {
+        return toArray(transform(failedToDeleteFiles.values(),
+                new Function<LocalIndexFile, String>() {
+                    @Override
+                    public String apply(LocalIndexFile input) {
+                        return input.deleteLog();
+                    }
+                }), String.class);
+    }
+
+    @Override
+    public String getGarbageSize() {
+        long garbageSize = 0;
+        for (LocalIndexFile failedToDeleteFile : failedToDeleteFiles.values()){
+            garbageSize += failedToDeleteFile.size;
+        }
+        return IOUtils.humanReadableByteCount(garbageSize);
+    }
+
+    @Override
+    public int getScheduledForCopyCount() {
+        return scheduledForCopyCount.get();
+    }
+
+    @Override
+    public int getCopyInProgressCount() {
+        return copyInProgressCount.get();
+    }
+
+    @Override
+    public String getCopyInProgressSize() {
+        return IOUtils.humanReadableByteCount(copyInProgressSize.get());
+    }
+
+    @Override
+    public int getMaxCopyInProgressCount() {
+        return maxCopyInProgressCount.get();
+    }
+
+    @Override
+    public int getMaxScheduledForCopyCount() {
+        return maxScheduledForCopyCount.get();
+    }
+
+    @Override
+    public String[] getCopyInProgressDetails() {
+        return toArray(transform(copyInProgressFiles,
+                new Function<LocalIndexFile, String>() {
+                    @Override
+                    public String apply(LocalIndexFile input) {
+                        return input.copyLog();
+                    }
+                }), String.class);
+    }
+
+    @Override
+    public int getDeletedFilesCount() {
+        return deletedFileCount.get();
+    }
+
+    @Override
+    public String getGarbageCollectedSize() {
+        return IOUtils.humanReadableByteCount(garbageCollectedSize.get());
+    }
+
     private static class IndexMappingData {
         static final String[] FIELD_NAMES = new String[]{
                 "jcrPath",
                 "fsPath",
+                "size",
         };
 
         static final String[] FIELD_DESCRIPTIONS = new String[]{
                 "JCR Path",
                 "Filesystem Path",
+                "Size",
         };
 
         static final OpenType[] FIELD_TYPES = new OpenType[]{
+                SimpleType.STRING,
                 SimpleType.STRING,
                 SimpleType.STRING,
         };
